Added function to add the save handler.
Tests don't work as the code for vtkExportConfiguration is not yet in the
add-in, and accessing code from another workbook which is not an
add-in is difficult.

That is to say that calling VBAToolKit_DEV.vtkExportConfiguration raises
an error 424 if we are not in VBAToolKit_DEV.

TODO : add the functionality in vtkCreateProject, when
vtkExportConfiguration will be in the add-in.